prueba con traits de medio boleto agregados
